lang:  Get rid of *unused* LANG_EQUALIZER_EDIT_MODE and all deprecated strings
The documentation says to not delete things but that dates back to when
we were a lot laxer in what we would accept at runtime.

I want to make things even stricter, and add a checksum or hash of
the string IDs, so that *any* change to the base language invalidates
translation and voice binaries at runtime.

Change-Id: I94a71781f839bec804dc04dd99660e5e866577bf