Improve the "No partition found" behavior.
(This is a different implementation of the fix in g#5726, by Vencislav Atanasov)

The core problem is that the user is prompted to insert a USB cable
to fix the partitioning etc but the code that monitors for USB insertion
hasn't been started yet.  Correct this.

If no USB support is present, reboot after 5 seconds if it's not a debug
build.

If USB support is present, want for insertion first, then do the
reboot-if-not-debug behavior.

Change-Id: I87827e7fe2fe9a02298918c6ebc4d8a9fb33d624