ATA:  Fix regression when trying to set power management mode with CF cards
Some CF cards claim to support Advanced Power Management, but error
out when the command is issued.  There is a special case in the
code that ignored errors when issuing the APM command, but was referenced
against a specific index in the feature table.

When the tagle was reorganized, the index was wrong depending
on if ATA_DMA was enabled or not.

The fix is to test against the specific subcommand ID instead of the
table index.

Change-Id: I2ef7a05b2d70675c38e41d70b5189b394056cb74