ErosQNative: Re-enable OF recovery
This is needed to be able to do updates to the OF due to the fact
that the OF reboots when you select "update firmware".

By default, most OF update images do not update the bootloader,
but it is easy to modify them to do so, and they already contain
a uboot.bin binary.

I tested using the Aigo Recovery from our native bootloader and it
sure seems to work fine.

If we patch the OF update images to load our own native bootloader,
this will be doubly necessary for uninstallation.

Change-Id: Id69b0321e293331b3936c3f3e7e012d5f00b479a