toolchain:  Make it easier to build toolchains on modern MacOS
When building on MacOS:

 * greadlink instead of readlink
 * gsed instead of sed
 * $TMPDIR instead of a hardcoded /tmp
 * apply a macos-specific patch
 * additional option to GCC

Additionally, don't rely on the system 'isl' stuff; always have binutils
build its own.

The forum post also says that bumping mpfr 3.1.x to 4.1.x, and mpc
from 1.0.x to 1.2.x is needed, but I do not want to do that due to the
possibility of unknown codegen changes.

Change-Id: I18084d392fa927318cc2a1183da7eb2ffa0d5884