S5L87xx bootloader: Add development menu (disabled by default)
A new flag S5L87XX_DEVELOPMENT_BOOTLOADER is introduced, to replace some of the bootloader functionality with debug/development menu. Not intended for end users, so it's not enabled unless explicitly specified when building

No changes to the binary if the new flag is not enabled.

Bootloader with the flag enabled boots on ipod6g.

This is a part of the large iPod Nano 3G and iPod Nano 4G support patch.

Change-Id: I3f28c94b4c1e29f3635285c7f0fd24a622a80c54