Revert back to the default 8px sysfont for ipod5g/ipod6g/erosq
The default status bar is sized for SYSFONT, so this made the status bar
slightly taller.

Unfortunately this broke a bunch of themes that didn't include a custom
status bar and had hardcoded assumptions about the height of the default.

While those themes should not be making those assumptions, we don't
yet have a straightforward way to deal with it.  So while we try to figure
out a better path forward, revert this.

Change-Id: I2cd6f64be5f405906a097ee0a33391e03bf1804d