Rename 'Playlist Catalogue' to 'Playlists'
When the Playlist Catalogue first appeared in the main menu
in 2011 (FS#11808), the 'Playlists' menu item was renamed
to 'Playlist Catalogue' which seems more verbose than is
necessary these days.

Since we don't have a 'Playlist Catalogue' menu item in the
onplay menu anymore, either, LANG_CATALOG can be eliminated.

Change-Id: I7f38b0847e9b982506371f4a4b473d1fb2551ad0