sb_create_from_settings: Fix crash when ui viewport setting malformed
The 'Purple_Glow_V3' theme uses the following
'UI viewport' setting which is missing a comma
after the viewport height:

ui viewport: 0,0,271,212-,-,-

The setting is actually ignored when loading this theme,
since a UI viewport is already defined in the supplied
SBS file.

Rockbox will crash, though, when and if the next theme
that is loaded resets the SBS by setting it to "-", but
doesn't adjust the UI viewport setting at the same time.

To fix, use default UI viewport when encountering fewer
parameters than expected for the UI viewport setting.

Change-Id: I2648565ac59369ce8ab18d8e84b7fa69ad767f25