languages:  Correct Latvian time reporting
12b9419006d stripped out leading/trailing whitespace, but the Latvian
translation used a blank string for VOICE_OH which got wiped out.

This string is only used when voicing the time, eg: 12:05 is spoken in
English as "twelve oh five".  Doing a little digging into Latvian, it
seems that for 24 hour clocks it would be voiced as "<twelve> nulle <five>"
instead, so I changed VOICE_OH to that.

Now for 12-hour clocks, it seems Latvian would voice "12:05" as "<five
past twelve>" but we don't currently have the ability to voice things
in this manner.  Will be addressed in a future patch.

Change-Id: I2c1bcd2e9caff508de3199da29c2e994e9d00c3c