disk:  GPT lookup code now supports virtual logical sector sizes
GPT superblocks are located at sector 1 and max_sector-1. If the system
uses a "virtual" sector that's larger than the drive's logical sector,
we need to map those virtual sector numbers to the drive's logical
sector.

If DEFAULT_VIRT_SECTOR_SIZE is defined, try that multiplier as well
as the standard multiplier of 1.

It's not practical to try every intermediate value so instead, if
DEFAULT_VIRT_SECTOR_SIZE is defined, try that as well as the standard
multiplier of 1.

This still leaves a handful of targets that don't set DEFAULT_VIRT
but do set MAX_VIRT.

Change-Id: I3accffcb97436b043836e072bfc620318a9b1230