misc: Suppress -Wformat-security when building with GCC7
Making splash -> splashf means its arguments are now checked
at compile time, but the "format" is nearly
always one of our virtual pointers instead of a string/format
literal.

Our gcc494 (and upcoming gcc950) toolchains handle this fine,
but simulator builds with gcc7 complain about this, so suppress
the warning for those environments.

(GCC 12, GCC 14, and GCC15 all seem to be okay with this too)

Change-Id: Ifaf061f14e2552db73a7515f61950ad83116e8b5